requirements.txt
asgiref==3.4.1
bcrypt==3.2.0
cffi==1.15.0
click==8.0.1
colorama==0.4.4
cryptography==35.0.0
ecdsa==0.17.0
fastapi==0.68.1
gunicorn==20.1.0
h11==0.12.0
mysql-connector-python==8.0.21
passlib==1.7.4
psycopg2==2.8.6
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.2
PyJWT==2.3.0
python-jose==3.3.0
python-multipart==0.0.5
rsa==4.7.2
six==1.16.0
starlette==0.14.2
typing-extensions==3.10.0.2
uvicorn==0.15.0